refactor: S3ProxyUploader 패키지 변경
